+++ packages/web-core/src/services/inworld_connection.service.ts
@@ -25,6 +25,7 @@ import { EventFactory } from '../factories/event';
 import { characterHasValidFormat, sceneHasValidFormat } from '../guard/scene';
 import { ConnectionService } from './connection.service';
 import { ConversationService } from './conversation.service';
+import { EntityService } from './entity.service';
 import { FeedbackService } from './feedback.service';
 
 interface InworldConnectionServiceProps<
@@ -40,6 +41,8 @@ export class InworldConnectionService<
   InworldPacketT extends InworldPacket = InworldPacket,
 > {
   readonly feedback: FeedbackService<InworldPacketT>;
+  readonly entity: EntityService<InworldPacketT>;
+
   private connection: ConnectionService<InworldPacketT>;
   private grpcAudioPlayer: GrpcAudioPlayback<InworldPacketT>;
   private oneToOneConversation: ConversationService<InworldPacketT>;
@@ -51,6 +54,7 @@ export class InworldConnectionService<
     this.connection = props.connection;
     this.grpcAudioPlayer = props.grpcAudioPlayer;
     this.feedback = new FeedbackService(props.connection);
+    this.entity = new EntityService(props.connection);
 
     this.player = new InworldPlayer<InworldPacketT>({
       grpcAudioPlayer: this.grpcAudioPlayer,